messages d'erreur si zone mal créée et ajout d'un bt pour créer une zone sur l'interface pour créer une zone

+++ src/application/view/EditionCreationZonesController.java
@@ -11,6 +11,8 @@ import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
@@ -32,52 +34,58 @@ public class EditionCreationZonesController {
 	public static Stage primaryStage;
 	private static Zone zoneAModifier;
 	private Zone sauvegardeZone;
-	
+
 	@FXML
 	VBox vbox;
-	
+
 	@FXML
 	ListView<ItemListPoint> points;
-	
+
 	@FXML
 	AnchorPane paneImage;
-	
+
+	@FXML
+	Button btNewZone;
+
+	@FXML
+	Label txtErreur;
+
 	@FXML
 	public void initialize() {
 		sauvegardeZone = new Zone(zoneAModifier.getIndex());
 		sauvegardeZone.getPoints().addAll(zoneAModifier.getPoints());
-		System.out.println(paneImage.getPrefWidth()+";"+paneImage.getPrefHeight());
-		System.out.println("img "+editionQuestionsZonesController.getThemeAModifier().getImageFond().getWidth()+";"+editionQuestionsZonesController.getThemeAModifier().getImageFond().getHeight());
-		double height=paneImage.getPrefHeight(),width=paneImage.getPrefWidth(), rapport=height/width;
-		if(editionQuestionsZonesController.getThemeAModifier().getImageFond().getWidth()*rapport>editionQuestionsZonesController.getThemeAModifier().getImageFond().getHeight())
-		{
-			height = width*editionQuestionsZonesController.getThemeAModifier().getImageFond().getHeight()/
-					editionQuestionsZonesController.getThemeAModifier().getImageFond().getWidth();
+		System.out.println(paneImage.getPrefWidth() + ";" + paneImage.getPrefHeight());
+		System.out.println("img " + editionQuestionsZonesController.getThemeAModifier().getImageFond().getWidth() + ";"
+				+ editionQuestionsZonesController.getThemeAModifier().getImageFond().getHeight());
+		double height = paneImage.getPrefHeight(), width = paneImage.getPrefWidth(), rapport = height / width;
+		if (editionQuestionsZonesController.getThemeAModifier().getImageFond().getWidth()
+				* rapport > editionQuestionsZonesController.getThemeAModifier().getImageFond().getHeight()) {
+			height = width * editionQuestionsZonesController.getThemeAModifier().getImageFond().getHeight()
+					/ editionQuestionsZonesController.getThemeAModifier().getImageFond().getWidth();
 			paneImage.setPrefHeight(height);
-		}
-		else
-		{
-			width = height*editionQuestionsZonesController.getThemeAModifier().getImageFond().getWidth()/
-					editionQuestionsZonesController.getThemeAModifier().getImageFond().getHeight();
+		} else {
+			width = height * editionQuestionsZonesController.getThemeAModifier().getImageFond().getWidth()
+					/ editionQuestionsZonesController.getThemeAModifier().getImageFond().getHeight();
 			paneImage.setPrefWidth(width);
 		}
-		BackgroundImage bgImage = new BackgroundImage(editionQuestionsZonesController.getThemeAModifier().getImageFond(),
-				BackgroundRepeat.NO_REPEAT, BackgroundRepeat.NO_REPEAT, BackgroundPosition.CENTER, new BackgroundSize(width, height, false, false, false, true));
+		BackgroundImage bgImage = new BackgroundImage(
+				editionQuestionsZonesController.getThemeAModifier().getImageFond(), BackgroundRepeat.NO_REPEAT,
+				BackgroundRepeat.NO_REPEAT, BackgroundPosition.CENTER,
+				new BackgroundSize(width, height, false, false, false, true));
 		paneImage.setBackground(new Background(bgImage));
 		zoneAModifier.setOpacity(0.3);
 		zoneAModifier.setFill(Color.RED);
 		zoneAModifier.setStroke(Color.BLACK);
 		paneImage.getChildren().add(zoneAModifier);
-		
+
 		Double x = null;
 		boolean xTraite = true;
-		for(Double coord : zoneAModifier.getPoints()) {
-			if(xTraite) {
-				x=coord;
-				xTraite=false;
-			}
-			else {
-				ItemListPoint itemPoint = new ItemListPoint(x,coord);
+		for (Double coord : zoneAModifier.getPoints()) {
+			if (xTraite) {
+				x = coord;
+				xTraite = false;
+			} else {
+				ItemListPoint itemPoint = new ItemListPoint(x, coord);
 				points.getItems().add(itemPoint);
 				itemPoint.setOnMouseClicked(event -> selectionPoint(itemPoint));
 				itemPoint.getCroix().setOnMouseClicked(event -> supprimerPoint(itemPoint));
@@ -85,67 +93,76 @@ public class EditionCreationZonesController {
 				paneImage.getChildren().add(point);
 				point.setTranslateX(x);
 				point.setTranslateY(coord);
-				xTraite=true;
+				xTraite = true;
 			}
 		}
-		
-		System.out.println(paneImage.getPrefWidth()+";"+paneImage.getPrefHeight());
-		Scaler.updateSize(Main.width,vbox);
-		System.out.println(paneImage.getPrefWidth()+";"+paneImage.getPrefHeight());
+
+		System.out.println(paneImage.getPrefWidth() + ";" + paneImage.getPrefHeight());
+		Scaler.updateSize(Main.width, vbox);
+		System.out.println(paneImage.getPrefWidth() + ";" + paneImage.getPrefHeight());
 	}
 
 	@FXML
 	private void addPoint(MouseEvent e) {
 		Circle point = new Circle(3, Color.RED);
 		paneImage.getChildren().add(point);
-		double x,y;
-		x=(double)(Math.round(e.getX()*100))/100;
-		y=(double)(Math.round(e.getY()*100))/100;
-		zoneAModifier.getPoints().add(x);
-		zoneAModifier.getPoints().add(y);
-		point.setTranslateX(x);
-		point.setTranslateY(y);
-		ItemListPoint itemPoint = new ItemListPoint(x,y);
-		itemPoint.setOnMouseClicked(event -> selectionPoint(itemPoint));
-		itemPoint.getCroix().setOnMouseClicked(event -> supprimerPoint(itemPoint));
-		points.getItems().add(itemPoint);
-		Scaler.updateSize(Main.width,itemPoint);
+		double x, y;
+		x = (double) (Math.round(e.getX() * 100)) / 100;
+		y = (double) (Math.round(e.getY() * 100)) / 100;
+		boolean alreadyExists = false;
+		for (ItemListPoint itemList : points.getItems())
+			if (itemList.getX() == x && itemList.getY() == y)
+				alreadyExists = true;
+		if (!alreadyExists) {
+			zoneAModifier.getPoints().add(x);
+			zoneAModifier.getPoints().add(y);
+			point.setTranslateX(x);
+			point.setTranslateY(y);
+			ItemListPoint itemPoint = new ItemListPoint(x, y);
+			itemPoint.setOnMouseClicked(event -> selectionPoint(itemPoint));
+			itemPoint.getCroix().setOnMouseClicked(event -> supprimerPoint(itemPoint));
+			points.getItems().add(itemPoint);
+			Scaler.updateSize(Main.width, itemPoint);
+		} else {
+			txtErreur.setText("Ce point a déjà été ajouté.");
+		}
 	}
-	
+
 	@FXML
 	public void supprimerPoint(ItemListPoint point) {
-		Circle circle=null;
-		for(Node node : paneImage.getChildren()) {
+		Circle circle = null;
+		for (Node node : paneImage.getChildren()) {
 			System.out.println(node);
-			if(node instanceof Circle) {
-				Circle circleTmp = (Circle)node;
-				System.out.println(circleTmp.getTranslateX()+";"+circleTmp.getTranslateY());
-				System.out.println(circleTmp.getCenterX()+";"+circleTmp.getCenterY());
-				if(circleTmp.getTranslateX()==point.getX().doubleValue() && circleTmp.getTranslateY()==point.getY().doubleValue()) {
-					circle = (Circle)node;
+			if (node instanceof Circle) {
+				Circle circleTmp = (Circle) node;
+				System.out.println(circleTmp.getTranslateX() + ";" + circleTmp.getTranslateY());
+				System.out.println(circleTmp.getCenterX() + ";" + circleTmp.getCenterY());
+				if (circleTmp.getTranslateX() == point.getX().doubleValue()
+						&& circleTmp.getTranslateY() == point.getY().doubleValue()) {
+					circle = (Circle) node;
 				}
 			}
 		}
 		points.getItems().remove(point);
 		zoneAModifier.getPoints().remove(point.getX());
 		zoneAModifier.getPoints().remove(point.getY());
-		if(circle!=null)
+		if (circle != null)
 			paneImage.getChildren().remove(circle);
 	}
-	
+
 	@FXML
 	public void selectionPoint(ItemListPoint point) {
 		System.out.println("clique on pt");
-		for(Node node : paneImage.getChildren()) {
+		for (Node node : paneImage.getChildren()) {
 			System.out.println(node);
-			if(node instanceof Circle) {
-				Circle circleTmp = (Circle)node;
-				System.out.println(circleTmp.getTranslateX()+";"+circleTmp.getTranslateY());
-				System.out.println(circleTmp.getCenterX()+";"+circleTmp.getCenterY());
-				if(circleTmp.getTranslateX()==point.getX().doubleValue() && circleTmp.getTranslateY()==point.getY().doubleValue()) {
+			if (node instanceof Circle) {
+				Circle circleTmp = (Circle) node;
+				System.out.println(circleTmp.getTranslateX() + ";" + circleTmp.getTranslateY());
+				System.out.println(circleTmp.getCenterX() + ";" + circleTmp.getCenterY());
+				if (circleTmp.getTranslateX() == point.getX().doubleValue()
+						&& circleTmp.getTranslateY() == point.getY().doubleValue()) {
 					circleTmp.setFill(Color.BLUE);
-				}
-				else {
+				} else {
 					circleTmp.setFill(Color.RED);
 				}
 			}
@@ -158,22 +175,46 @@ public class EditionCreationZonesController {
 
 	@FXML
 	public void valider() throws IOException {
-		VBox root = new VBox();
+		if (points.getItems().size() > 2) {
+			VBox root = new VBox();
 
-		editionQuestionsZonesController.primaryStage = primaryStage;
-		root = FXMLLoader.load(getClass().getResource("editionQuestionsZones.fxml"));
-		Scene scene = new Scene(root);
+			editionQuestionsZonesController.primaryStage = primaryStage;
+			root = FXMLLoader.load(getClass().getResource("editionQuestionsZones.fxml"));
+			Scene scene = new Scene(root);
 
-		primaryStage.setResizable(false);
+			primaryStage.setResizable(false);
 
-		primaryStage.setScene(scene);
+			primaryStage.setScene(scene);
+		} else
+			txtErreur.setText("Il faut que la zone contiennent au moins trois points.");
+	}
+
+	@FXML
+	public void creerNewZone() throws IOException {
+		if (points.getItems().size() > 2) {
+			Zone newZ = new Zone(editionQuestionsZonesController.idZoneMax);
+			editionQuestionsZonesController.idZoneMax++;
+			editionQuestionsZonesController.getThemeAModifier().addZone(newZ);
+			EditionCreationZonesController.setZone(newZ);
+			try {
+				VBox root = null;
+				root = FXMLLoader.load(getClass().getResource("Creation Zone.fxml"));
+				Scene scene = new Scene(root, Main.width, Main.height);
+				primaryStage.setResizable(false);
+				primaryStage.setScene(scene);
+				primaryStage.show();
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+		} else
+			txtErreur.setText("Il faut que la zone contiennent au moins trois points.");
 	}
 
 	@FXML
 	public void quitter() throws IOException {
 		zoneAModifier.getPoints().clear();
 		zoneAModifier.getPoints().addAll(sauvegardeZone.getPoints());
-		if(sauvegardeZone.getPoints().size()<3)
+		if (sauvegardeZone.getPoints().size() < 3)
 			editionQuestionsZonesController.getThemeAModifier().getZones().remove(zoneAModifier);
 		VBox root = new VBox();
